Extend Node drain e2e test for MachineDrainRules
